+++ backend/src/main/java/wooteco/prolog/login/ui/auto_interceptor_uri_mapper/scanner/ControllerScanner.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login.ui.auto_interceptor_register.scanner;
+package wooteco.prolog.login.ui.auto_interceptor_uri_mapper.scanner;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RestController;
