--- backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_register/test_classes/NormalClass.java
@@ -1,4 +1,0 @@
-package wooteco.prolog.login.ui.auto_interceptor_register.test_classes;
-
-public class NormalClass {
-}
