+++ backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_uri_mapper/test_classes/ControllerClass.java
@@ -1,9 +1,8 @@
-package wooteco.prolog.login.ui.auto_interceptor_register.test_classes;
+package wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 import wooteco.prolog.login.domain.AuthMemberPrincipal;
 import wooteco.prolog.member.domain.Member;
