+++ backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_uri_mapper/test_classes/RestControllerClass.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login.ui.auto_interceptor_register.test_classes;
+package wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
